kernel: Pull out avplayer,auidodev,videodec,fiber state kernel: Add ObjectStore modules: Support library init function modules: Pull out cpu_protocol vita3k: Rename .h -> .inc io: Remove unneccesary base_tick argument cmake: Folderify modules and kernel
