+++ vita3k/host/CMakeLists.txt
@@ -3,8 +3,6 @@ add_library(
 	STATIC
 	include/host/app_util.h
 	include/host/functions.h
-	include/host/import_var.h
-	include/host/import_fn.h
 	include/host/load_self.h
 	include/host/pkg.h
 	include/host/sce_types.h
@@ -20,5 +18,5 @@ add_library(
 )
 
 target_include_directories(host PUBLIC include ${PSVPFSPARSER_INCLUDE_DIR})
-target_link_libraries(host PUBLIC psvpfsparser app audio config ctrl dialog ime io kernel miniz net ngs nids np renderer sdl2 touch gdbstub)
+target_link_libraries(host PUBLIC psvpfsparser app audio config ctrl dialog ime io kernel miniz net ngs nids np renderer sdl2 touch gdbstub codec)
 target_link_libraries(host PRIVATE elfio::elfio FAT16 vita-toolchain)
