+++ vita3k/modules/SceAvPlayer/SceAvPlayer.cpp
@@ -17,6 +17,7 @@
 
 #include "SceAvPlayer.h"
 
+#include <codec/state.h>
 #include <io/functions.h>
 #include <kernel/thread/thread_functions.h>
 #include <util/lock_and_find.h>
@@ -39,6 +40,65 @@ typedef int (*SceAvPlayerReadFile)(void *arguments, uint8_t *buffer, uint64_t of
 typedef uint64_t (*SceAvPlayerGetFileSize)(void *arguments);
 
 typedef void (*SceAvPlayerEventCallback)(void *arguments, int32_t event_id, int32_t source_id, void *event_data);
+
+struct PlayerInfoState;
+typedef std::shared_ptr<PlayerInfoState> PlayerPtr;
+typedef std::map<SceUID, PlayerPtr> PlayerStates;
+
+struct AvPlayerState {
+    std::mutex mutex;
+    PlayerStates players;
+};
+
+struct SceAvPlayerMemoryAllocator {
+    uint32_t user_data;
+
+    // All of these should be cast to SceAvPlayerAllocator or SceAvPlayerDeallocator types.
+    Ptr<void> general_allocator;
+    Ptr<void> general_deallocator;
+    Ptr<void> texture_allocator;
+    Ptr<void> texture_deallocator;
+};
+
+struct SceAvPlayerFileManager {
+    uint32_t user_data;
+
+    // Cast to SceAvPlayerOpenFile, SceAvPlayerCloseFile, SceAvPlayerReadFile and SceAvPlayerGetFileSize.
+    Ptr<void> open_file;
+    Ptr<void> close_file;
+    Ptr<void> read_file;
+    Ptr<void> file_size;
+};
+
+struct SceAvPlayerEventManager {
+    uint32_t user_data;
+
+    // Cast to SceAvPlayerEventCallback.
+    Ptr<void> event_callback;
+};
+
+struct PlayerInfoState {
+    PlayerState player;
+
+    // Framebuffer count is defined in info. I'm being safe now and forcing it to 4 (even though its usually 2).
+    constexpr static uint32_t RING_BUFFER_COUNT = 4;
+
+    uint32_t video_buffer_ring_index = 0;
+    uint32_t video_buffer_size = 0;
+    std::array<Ptr<uint8_t>, RING_BUFFER_COUNT> video_buffer;
+
+    uint32_t audio_buffer_ring_index = 0;
+    uint32_t audio_buffer_size = 0;
+    std::array<Ptr<uint8_t>, RING_BUFFER_COUNT> audio_buffer;
+
+    bool do_loop = false;
+    bool paused = false;
+
+    uint64_t last_frame_time = 0;
+    SceAvPlayerMemoryAllocator memory_allocator;
+    SceAvPlayerFileManager file_manager;
+    SceAvPlayerEventManager event_manager;
+};
 
 enum class DebugLevel {
     NONE,
@@ -175,7 +235,8 @@ uint run_event_callback(HostState &host, SceUID thread_id, const PlayerPtr playe
 //end of callback_thread
 
 EXPORT(int32_t, sceAvPlayerAddSource, SceUID player_handle, Ptr<const char> path) {
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, state->mutex);
 
     auto file_path = expand_path(host.io, path.get(host.mem), host.pref_path);
     if (!fs::exists(file_path) && player_info->file_manager.open_file && player_info->file_manager.close_file && player_info->file_manager.read_file && player_info->file_manager.file_size) {
@@ -220,14 +281,17 @@ EXPORT(int32_t, sceAvPlayerAddSource, SceUID player_handle, Ptr<const char> path
 }
 
 EXPORT(int, sceAvPlayerClose, SceUID player_handle) {
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, state->mutex);
     run_event_callback(host, thread_id, player_info, SCE_AVPLAYER_STATE_STOP, 0, Ptr<void>(0));
-    host.kernel.players.erase(player_handle);
+    std::lock_guard<std::mutex> lock(state->mutex);
+    state->players.erase(player_handle);
     return 0;
 }
 
 EXPORT(uint64_t, sceAvPlayerCurrentTime, SceUID player_handle) {
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, state->mutex);
 
     return player_info->player.last_timestamp;
 }
@@ -252,7 +316,8 @@ EXPORT(int32_t, sceAvPlayerEnableStream, SceUID player_handle, uint32_t stream_n
 }
 
 EXPORT(bool, sceAvPlayerGetAudioData, SceUID player_handle, SceAvPlayerFrameInfo *frame_info) {
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, state->mutex);
 
     Ptr<uint8_t> buffer;
 
@@ -292,7 +357,8 @@ EXPORT(uint32_t, sceAvPlayerGetStreamInfo, SceUID player_handle, uint stream_no,
         return SCE_AVPLAYER_ERROR_ILLEGAL_ADDR;
     }
     STUBBED("ALWAYS SUSPECTS 2 STREAMS: VIDEO AND AUDIO");
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, state->mutex);
     auto StreamInfo = stream_info.get(host.mem);
     if (stream_no == 0) { //suspect always two streams: audio and video //first is video
         DecoderSize size = player_info->player.get_size();
@@ -315,7 +381,8 @@ EXPORT(uint32_t, sceAvPlayerGetStreamInfo, SceUID player_handle, uint stream_no,
 }
 
 EXPORT(bool, sceAvPlayerGetVideoData, SceUID player_handle, SceAvPlayerFrameInfo *frame_info) {
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, state->mutex);
 
     Ptr<uint8_t> buffer;
 
@@ -363,10 +430,12 @@ EXPORT(bool, sceAvPlayerGetVideoDataEx, SceUID player_handle, SceAvPlayerFrameIn
 }
 
 EXPORT(SceUID, sceAvPlayerInit, SceAvPlayerInfo *info) {
+    host.kernel.obj_store.create<AvPlayerState>();
     if (host.cfg.current_config.video_playing) {
+        const auto state = host.kernel.obj_store.get<AvPlayerState>();
         SceUID player_handle = host.kernel.get_next_uid();
         PlayerPtr player = std::make_shared<PlayerInfoState>();
-        host.kernel.players[player_handle] = player;
+        state->players[player_handle] = player;
 
         LOG_TRACE("SceAvPlayerInfo.memory_allocator: user_data:{}, general_allocator:{}, general_deallocator:{}, texture_allocator:{}, texture_deallocator:{}",
             log_hex(info->memory_allocator.user_data), log_hex(info->memory_allocator.general_allocator.address()), log_hex(info->memory_allocator.general_deallocator.address()),
@@ -393,7 +462,8 @@ EXPORT(SceUID, sceAvPlayerInit, SceAvPlayerInfo *info) {
 }
 
 EXPORT(bool, sceAvPlayerIsActive, SceUID player_handle) {
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, state->mutex);
 
     return !player_info->player.video_playing.empty();
 }
@@ -403,7 +473,8 @@ EXPORT(int, sceAvPlayerJumpToTime) {
 }
 
 EXPORT(int, sceAvPlayerPause, SceUID player_handle) {
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, state->mutex);
     player_info->paused = true;
     run_event_callback(host, thread_id, player_info, SCE_AVPLAYER_STATE_PAUSE, 0, Ptr<void>(0));
     return 0;
@@ -414,7 +485,8 @@ EXPORT(int, sceAvPlayerPostInit) {
 }
 
 EXPORT(int, sceAvPlayerResume, SceUID player_handle) {
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, state->mutex);
     if (!player_info->paused) {
         run_event_callback(host, thread_id, player_info, SCE_AVPLAYER_STATE_PLAY, 0, Ptr<void>(0));
     }
@@ -423,7 +495,8 @@ EXPORT(int, sceAvPlayerResume, SceUID player_handle) {
 }
 
 EXPORT(int, sceAvPlayerSetLooping, SceUID player_handle, bool do_loop) {
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, state->mutex);
     player_info->do_loop = do_loop;
 
     return STUBBED("LOOPING NOT IMPLEMENTED");
@@ -434,7 +507,8 @@ EXPORT(int, sceAvPlayerSetTrickSpeed) {
 }
 
 EXPORT(int, sceAvPlayerStart, SceUID player_handle) {
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, state->mutex);
     if (!player_info->player.videos_queue.empty()) {
         player_info->player.pop_video();
     }
@@ -443,7 +517,8 @@ EXPORT(int, sceAvPlayerStart, SceUID player_handle) {
 }
 
 EXPORT(int, sceAvPlayerStop, SceUID player_handle) {
-    const PlayerPtr &player_info = lock_and_find(player_handle, host.kernel.players, host.kernel.mutex);
+    const auto state = host.kernel.obj_store.get<AvPlayerState>();
+    const PlayerPtr &player_info = lock_and_find(player_handle, state->players, host.kernel.mutex);
     player_info->player.free_video();
     run_event_callback(host, thread_id, player_info, SCE_AVPLAYER_STATE_STOP, 0, Ptr<void>(0));
     return 0;
