+++ vita3k/modules/include/modules/library_init_list.inc
@@ -1,0 +1,1 @@
+LIBRARY(SceFiber)
