+++ vita3k/modules/include/modules/module_parent.h
@@ -24,8 +24,8 @@
 struct CPUState;
 struct HostState;
 struct KernelState;
-struct CPUDepInject;
 
+void init_libraries(HostState &host);
 void call_import(HostState &host, CPUState &cpu, uint32_t nid, SceUID thread_id);
 bool load_module(HostState &host, SceSysmoduleModuleId module_id);
 Address resolve_export(KernelState &kernel, uint32_t nid);
@@ -41,22 +41,15 @@ struct VarExport {
 constexpr int var_exports_size =
 #define NID(name, nid)
 #define VAR_NID(name, nid) 1 +
-#include <nids/nids.h>
+#include <nids/nids.inc>
     0;
 #undef VAR_NID
 #undef NID
 
 const std::array<VarExport, var_exports_size> &get_var_exports();
 
-// TODO move to module
-struct CPUProtocol : public CPUProtocolBase {
-    CPUProtocol(HostState &host);
-    ~CPUProtocol();
-    void call_svc(CPUState &cpu, uint32_t svc, Address pc, SceUID thread_id) override;
-    Address get_watch_memory_addr(Address addr) override;
-    std::vector<ModuleRegion> &get_module_regions() override;
-    ExclusiveMonitorPtr get_exlusive_monitor() override;
+using LibraryInitFn = std::function<void(HostState &host)>;
 
-private:
-    HostState *host;
-};
+#define LIBRARY_INIT_DECL(name) void export_library_init_##name(HostState &host);
+#define LIBRARY_INIT_IMPL(name) void export_library_init_##name(HostState &host)
+#define LIBRARY_INIT_REGISTER(name) extern const LibraryInitFn import_library_init_##name = export_library_init_##name;
