+++ vita3k/nids/CMakeLists.txt
@@ -2,7 +2,7 @@ add_library(
 	nids
 	STATIC
 	include/nids/functions.h
-	include/nids/nids.h
+	include/nids/nids.inc
 	include/nids/types.h
 	src/nids.cpp
 )
