+++ vita3k/nids/src/nids.cpp
@@ -2,7 +2,7 @@
 
 #define VAR_NID(name, nid) extern const char name_##name[] = #name;
 #define NID(name, nid) extern const char name_##name[] = #name;
-#include <nids/nids.h>
+#include <nids/nids.inc>
 #undef NID
 #undef VAR_NID
 
@@ -14,7 +14,7 @@ const char *import_name(uint32_t nid) {
 #define NID(name, nid) \
     case nid:          \
         return name_##name;
-#include <nids/nids.h>
+#include <nids/nids.inc>
 #undef NID
 #undef VAR_NID
     default:
